raxy_project/Farm-046fa22dd30478ab4290675caa75cf9e01a0cc8f/requirements.txt
botasaurus
rich
supabase
random-user-agent
flask
fastapi
uvicorn[standard]
typer[all]
loguru
dependency-injector
pyyaml
httpx
python-dotenv
urllib3
redis
